+++ model/parts/system_metrics.py
@@ -153,6 +153,9 @@ def policy_total_online_validator_rewards(
     total_priority_fee_to_validators = previous_state[
         "total_priority_fee_to_validators"
     ]
+    total_realized_mev_to_validators = previous_state[
+        "total_realized_mev_to_validators"
+    ]
 
     # Calculate total rewards for online validators
     total_online_validator_rewards = (
@@ -160,6 +163,7 @@ def policy_total_online_validator_rewards(
         - validating_penalties
         + whistleblower_rewards
         + total_priority_fee_to_validators
+        + total_realized_mev_to_validators * constants.gwei
     )
 
     return {"total_online_validator_rewards": total_online_validator_rewards}