+++ model/state_variables.py
@@ -51,7 +51,7 @@ class StateVariables:
     """
     The stage of the network upgrade process.
 
-    By default set to PROOF_OF_STAKE Stage, where EIP1559 is enabled and POW issuance is disabled.
+    By default set to PROOF_OF_STAKE Stage, where EIP-1559 is enabled and POW issuance is disabled.
 
     Otherwise set to ALL Stage, which transitions through each stage, updating the `stage` State Variable.
 
@@ -119,15 +119,21 @@ class StateVariables:
     whistleblower_rewards: Gwei = 0
     """The total rewards received as a proportion of the effective balance of the slashed validators"""
 
-    # EIP1559 state variables
+    # EIP-1559 state variables
     base_fee_per_gas: Gwei_per_Gas = 1
     """The base fee burned, in Gwei per gas, dynamically updated for each block"""
     total_base_fee: Gwei = 0
     """The total base fee burned for all transactions included in blockspace"""
     total_priority_fee_to_miners: Gwei = 0
-    """"The total priority fee to miners pre-merge for all transactions included in blockspace"""
+    """"The total priority fee to miners pre Proof-of-Stake for all transactions included in blockspace"""
     total_priority_fee_to_validators: Gwei = 0
-    """"The total priority fee to validators post-merge for all transactions included in blockspace"""
+    """"The total priority fee to validators post Proof-of-Stake for all transactions included in blockspace"""
+
+    # Maximum Extractable Value (MEV) state variables
+    total_realized_mev_to_miners: ETH = 0
+    """The total realized MEV to miners pre Proof-of-Stake"""
+    total_realized_mev_to_validators: ETH = 0
+    """The total realized MEV to validators post Proof-of-Stake"""
 
     # System metric state variables
     validator_eth_staked: np.ndarray = np.zeros(
@@ -152,7 +158,6 @@ class StateVariables:
     )
     """The total annualized profit (income received - costs) yields (percentage of investment amount)
     per validator environment"""
-
     validator_count_distribution: np.ndarray = np.zeros(
         (number_of_validator_environments, 1), dtype=int
     )