+++ model/system_parameters.py
@@ -32,7 +32,10 @@ from model.types import (
     Stage,
 )
 from model.utils import default
-from data.historical_values import eth_price_mean
+from data.historical_values import (
+    eth_price_mean,
+    eth_block_rewards_mean,
+)
 
 
 # Configure validator environment distribution
@@ -134,6 +137,8 @@ class Parameters:
     system parameter key: system parameter type = default system parameter value
 
     Because lists are mutable, we need to wrap each parameter list in the `default(...)` method.
+
+    For default value assumptions, see the ASSUMPTIONS.md document.
     """
 
     # Time parameters
@@ -152,7 +157,7 @@ class Parameters:
     """
     Which stage or stages of the network upgrade process to simulate.
 
-    By default set to PROOF_OF_STAKE stage, where EIP1559 is enabled and POW issuance is disabled.
+    By default set to PROOF_OF_STAKE stage, where EIP-1559 is enabled and POW issuance is disabled.
 
     See model.types.Stage Enum for further documentation.
     """
@@ -164,14 +169,12 @@ class Parameters:
         [datetime.strptime("2021/08/04", "%Y/%m/%d")]
     )
     """
-    Expected EIP1559 activation date as Python datetime.
+    Expected EIP-1559 activation date as Python datetime.
     """
 
     date_pos: List[datetime] = default([datetime.strptime("2021/12/1", "%Y/%m/%d")])
     """
     Eth1/Eth2 merge date as Python datetime, after which POW is disabled and POS is enabled.
-
-    Source: https://twitter.com/drakefjustin/status/1379052831982956547
     """
 
     # Environmental processes
@@ -203,7 +206,7 @@ class Parameters:
         ]
     )
     """
-    A process that returns the number of new validators per epoch.
+    A process that returns the number of new validators added to the activation queue per epoch.
 
     Used if model not driven using `eth_staked_process`.
 
@@ -211,11 +214,28 @@ class Parameters:
     """
 
     # Ethereum system parameters
-    daily_pow_issuance: List[ETH] = default([13_550])
+    daily_pow_issuance: List[ETH] = default([eth_block_rewards_mean])
     """
-    The average daily Proof of Work issuance in ETH.
+    The average daily Proof-of-Work issuance in ETH.
 
     See https://etherscan.io/chart/blockreward
+    """
+
+    mev_per_block: List[ETH] = default([0])
+    """
+    By default the realized Maximum Extractable Value (MEV) per block is set to zero
+    to only consider the influence of Proof-of-Stake (PoS) incentives on validator yields.
+    
+    To investigate the influence of MEV on validator yields,
+    set this parameter to a reasonable value for the realized MEV per block / slot.
+    
+    The realized MEV per block is allocated to miners pre-PoS and validators post-PoS,
+    increasing the effective yields of those miners / validators
+    that use MEV clients such as Flashbots' MEV-geth.
+    
+    An example of a valid assumption for the realized MEV
+    would be the 30-day realized MEV from https://explore.flashbots.net/,
+    this value can then be calculated per-block to set the `mev_per_block` parameter.
     """
 
     # Parameters from the Eth2 specification
@@ -300,11 +320,11 @@ class Parameters:
     """
     BASE_FEE_MAX_CHANGE_DENOMINATOR: List[int] = default([8])
     """
-    Used to set the maximum rate at which the EIP1559 base fee can change per block, approx. 12.5%.
+    Used to set the maximum rate at which the EIP-1559 base fee can change per block, approx. 12.5%.
     """
     ELASTICITY_MULTIPLIER: List[int] = default([2])
     """
-    Used to calculate gas limit from EIP1559 gas target
+    Used to calculate gas limit from EIP-1559 gas target
     """
 
     # Validator parameters
@@ -357,20 +377,18 @@ class Parameters:
     The number of slashing events per 1000 epochs.
     """
 
-    # EIP1559 transaction pricing parameters
+    # EIP-1559 transaction pricing parameters
     base_fee_process: List[Callable[[Run, Timestep], Gwei_per_Gas]] = default(
-        [lambda _run, _timestep: 70]  # Gwei per gas
+        [lambda _run, _timestep: 30]  # Gwei per gas
     )
     """
-    The base fee burned, in Gwei per gas, for each transaction.
+    EIP-1559 transaction pricing base fee burned, in Gwei per gas, for each transaction.
+    See https://eips.ethereum.org/EIPS/eip-1559 for EIP-1559 proposal.
 
-    An average of 100 Gwei per gas expected to be set as transaction fee cap,
-    split between the base fee and priority fee - the fee cap less the base fee is sent as a priority fee to miners/validators.
+    See ASSUMPTIONS.md doc for further details about default value assumption.
 
-    Approximated using average gas price from https://etherscan.io/gastracker as of 20/04/21
-
-    An extract from https://notes.ethereum.org/@vbuterin/eip-1559-faq
-
+    An extract from https://notes.ethereum.org/@vbuterin/eip-1559-faq:
+    
     > Each “full block” (ie. a block whose gas is 2x the TARGET) increases the BASEFEE by 1.125x,
     > so a series of constant full blocks will increase the gas price by a factor of 10 every
     > ~20 blocks (~4.3 min on average).
@@ -378,17 +396,13 @@ class Parameters:
     """
 
     priority_fee_process: List[Callable[[Run, Timestep], Gwei_per_Gas]] = default(
-        [lambda _run, _timestep: 30]  # Gwei per gas
+        [lambda _run, _timestep: 2]  # Gwei per gas
     )
     """
-    EIP1559 transaction pricing priority fee, in Gwei per gas.
-
-    Due to MEV, average priority fee expected to be higher than usual as bid for inclusion in blockscpace market.
-
-    The priority fee is the difference between the fee cap set per transaction, and the base fee.
-
-    For PoW system without MEV influence, the priority fee level compensates for uncle risk:
-    See https://notes.ethereum.org/@vbuterin/BkSQmQTS8#Why-would-miners-include-transactions-at-all
+    EIP-1559 transaction pricing average priority fee, in Gwei per gas, for each transaction.
+    See https://eips.ethereum.org/EIPS/eip-1559 for EIP-1559 proposal.
+    
+    See ASSUMPTIONS.md doc for further details about default value assumption.
     """
 
     gas_target_process: List[Callable[[Run, Timestep], Gas]] = default(
@@ -401,8 +415,9 @@ class Parameters:
     * a “long-term average target” (equal to the current gas limit) == gas target
     * a “hard per-block cap” (twice the current gas limit) == gas limit
 
-    EIP1559 gas limit = gas_target * ELASTICITY_MULTIPLIER
-    See https://eips.ethereum.org/EIPS/eip-1559
+    EIP-1559 gas limit = gas_target * ELASTICITY_MULTIPLIER
+    
+    See ASSUMPTIONS.md doc for further details about default value assumption.
     """